requirements.txt
django
djangorestframework
psycopg2-binary
python-decouple
pillow
drf-yasg
gunicorn
celery
redis
djangorestframework_simplejwt